Alarm handling: Fix pointer popup
Pointers weren't being displayed when alarm fired while app was in background
and the current view was other than the main view. Main view made the assumption
that the app would only fire while the main view was currently being displayed, and that
dismissing the pointer view would return to the main view.