Connection: Add refresh button for failed connection
Allow the user to manually trigger a refreshing of data. Refresh button
is made visible if app can't connect to the internet when it is loaded.
App will attempt to reconnect and load data from xml file once the user
hits the refresh button. If refresh succeeds, the refresh button is hidden.